Added Level system and Level changed event

+++ Assets/Scripts/World/RPG/LevelComp.cs
@@ -3,7 +3,7 @@
     public struct LevelComp
     {
         public int Level;
-        public int Experience;
-        public int ExperienceToNextLevel;
+        public float Experience;
+        public float ExperienceToNextLevel;
     }
 }
